+++ docs/release-notes.md
@@ -1,5 +1,17 @@
 # Release notes
 
+## Version 0.0.5
+
+### Added
+
+- `sorted_exports` for sorted `__all__` exports.
+- Hooks for add_lazy_import, add_deprecated_lazy_import.
+
+### Changed
+
+- `find_missing` test method has some different error names.
+- `find_missing` doesn't require the all_var anymore.
+
 ## Version 0.0.4
 
 ### Added